+++ src/__init__.py
@@ -5,7 +5,7 @@ import yaml
 from flask import Flask
 
 from src.db_model.db_models import db
-from src.schedule import schedule_bp, bus_bp, driver_bp, available_schedule_bp
+from src.blueprints import schedule_bp, bus_bp, driver_bp, available_schedule_bp
 from src.constants.http_status_codes import HTTP_200_OK
 
 coloredlogs.install(level='DEBUG')
